Hostname verification (#67)
* Add failing Test

* Fix Testnames in Cmake Config

* Set Servername for OpenSSL in Tests

Makes the wrong hostname test work for system that do not need the native verification

* Set Verification Callback via SSL_CTX_set_verify

Also remove some unnecessary functions in verification_utils.hpp

This fixes hostname verification on systems that use the native implementations

* Update Github Test Chain

Test failed with "self signed certificate in certificate chain in function"

* Silence Conversion Warning size_t->DWORD

* Add missing openssl includes

* Fix error in appveyor build

X509_STORE_get0_param was only introduced in OpenSSL1.1.0
The Appveyor build uses OpenSSL 1.0.2
Use X509_STORE_CTX_get0_param from within store_ctx instead

* Fix spki digest test

I previously updated the old (expired) Github cert chain,
inadvertently breaking the test.

Re-add it to the fail_chains and adapt the test report_errors.

+++ include/boost/certify/detail/keystore_windows.ipp
@@ -96,7 +96,7 @@ BOOST_CERTIFY_DECL std::unique_ptr<::CERT_CONTEXT const, cert_context_deleter>
     if (!CertAddEncodedCertificateToStore(store.get(),
                                           X509_ASN_ENCODING,
                                           buffer.data(),
-                                          buffer.size(),
+                                          static_cast<DWORD>(buffer.size()),
                                           CERT_STORE_ADD_ALWAYS,
                                           &leaf_cert))
         return nullptr;
@@ -110,7 +110,7 @@ BOOST_CERTIFY_DECL std::unique_ptr<::CERT_CONTEXT const, cert_context_deleter>
         if (!CertAddEncodedCertificateToStore(store.get(),
                                               X509_ASN_ENCODING,
                                               buffer.data(),
-                                              buffer.size(),
+                                              static_cast<DWORD>(buffer.size()),
                                               CERT_STORE_ADD_ALWAYS,
                                               nullptr))
             return nullptr;
